requirements.txt
Flask
numpy
Pillow
torch==2.3.1
torchvision==0.18.1
Werkzeug
